Metavm moved from Model to loader for Schema/Model browser compatibility
