+++ lib/loader.js
@@ -6,6 +6,7 @@ const path = require('path');
 const metavm = require('metavm');
 
 const { Schema } = require('./schema.js');
+const { Model } = require('./model.js');
 
 const createSchema = (name, src) => {
   const { exports } = new metavm.MetaScript(name, src);
@@ -34,4 +35,22 @@ const readDirectory = async (dirPath) => {
   return structs;
 };
 
-module.exports = { createSchema, loadSchema, readDirectory };
+const loadModel = async (modelPath, systemTypes = {}) => {
+  const structs = await readDirectory(modelPath);
+  const database = structs.get('.database') || null;
+  const customTypes = structs.get('.types') || {};
+  structs.delete('.database');
+  structs.delete('.types');
+  const types = { ...systemTypes, ...customTypes };
+  return new Model(types, structs, database);
+};
+
+const saveTypes = (outputFile, model) => fsp.writeFile(outputFile, model.dts);
+
+module.exports = {
+  createSchema,
+  loadSchema,
+  readDirectory,
+  loadModel,
+  saveTypes,
+};
